Add rootDirectory
This enables running writ over a directory different than the cwd.

In addition to this, the exportDirectory is joined together with the rootDirectory in order to be able to export the build also to the correct destination.

The change above means the user cannot specify an exportDirectory that is outside of the main directory.

A nuance to be aware of is that rootDirectory is an absolute path, while the exportDirectory is expected to be a relative path (to the root).