Preserve original newlines after live-edit
If a post is markdown and the author has used single newlines to format line lengths, later when the author uses live editor, those original line breaks will be preserved.

Also be more mindful about white-space usage in rendering.

Later, a need for trimming trailing white-spaces from the live-edited file content and it's also solved in this commit