Discourage watch mode usage as it is broken
How preview and expansions are implemented should be redesigned before confidently using watch mode. Currently, if more than a certain amount of assets are copied, it causes watcher to re-run infinitely due to duplicate init call within it.

To remove the duplicate call within the watcher, which is currently the only way to actually propagate preview expansions, preview expansions should be registered before compilation. That requires re-thinking that I won't, for now.