Allow pass through localAsset folders
These would be previously classified as unrecognized directory (and their content as unrecognized file).
Remove unrecognizedDirectory and unrecognizedFile in favor of localAsset folders.

Refactor the fsTree to contentModel mapping to cleanly separate deep branching of category and folderedPost from each other. Subsequently, a foldered post at the root can have a localAsset folder that has a post.md in it, without causing it to be registered as a non-sensical nested foldered post