Introduce mentions
Posts, subpages and categories can be mentioned by other posts, subpages and categories
A generic 'mention' templateHelper with is added. At simplest, it takes a permalink and returns a link. If used as a block helper, it will provide the mentioned entry as context

Recognize cover and media metadata in parseTemplate, so they can be always prefixed with the entry's permalink inside contentModel. This was especially necessary when an entry's cover or media is used in a mention. No need to prefix entry permalink when these are used outside of entry's own page