Use slug module to fix: #8, #10
The bug was due to the character code difference between when a file is created in vim vs in finder.
The slug module takes care of it. It also removes the extension and slash characters.
So, a bunch of changes were necessary to keep everything that relied on getSlug working.