Fix post mentioning by setting ctx to contentModel
Now that the post content template is aware that it is rendering a post content ({{#with post}}), subpage can't use it any more.
This means subpage having its own content template and a footer, while at it. But having classnames with the prefix 'post' does not make sense in subpage content and footer.
And that means, the frontend assets should be abstracted enough to handle both posts and subpages now (instead of both being able to use a single script)

Closes: #9