Sub-contentModels WIP
Mainly trying to render a basic portal with a homepage, a subpage and a blog

Attempt making portal to be the root content model that initiates
sub-content-models depending on fileSystemTree (WIP)

Rendering
Different content models will have different models with different target views.
Re-think how content models can be rendered. An idea was to mimic an MVC
architecture for content model so it can have views that contain specific rendering targets

parseTemplate seems it should be cleaned in favor of moving those blog-aware metadata to models.

read TODO comments in the code