[CM2] Introduce ContentTypes and root.md
Now standalone content types can be stored in a schema folder
(customizable via settings). For now, they have the content-modelling
properties + a model property which points to a model.  e.g.  collection, post

When a collection is matched with a content type, it inherits its
properties.

Also, collectionAliases can be put in a root.md to allow for more
default collectionIndexFile names. Other content-modelling properties in
root.md are skipped for now.

With this change, now, the collection.md itself can be named in a
domain-driven way. e.g. blog.md, portfolio.md, podcast.md

Also:

- Fix RegExpes that match indexFiles in models.
  #til: new RegExp(`\.`) and /\./ won't match the same thing.