[CM2] A collection can render at the root if empty slug
If 'slug:' is added to its frontMatter, its permalink and outputPath will collapse into those of its parent context.

When the parent permalink is just '/', child permalinks would break. So, a simple makePermalink helper is used to encapsulate dealing with this case. This helper is used in all models, with the idea of turning any page into a homepage by using the slug: technique. But this is not implemented in models beyond collection, as it'd take more (and ugly) work.

If turning any entry or index page into a homepage using the slug technique worked easily, it could work behind the scenes of a 'homepage' frontMatter property. But maybe later.

Anyway, rendering a collection in the root is especially helpful for sites with just one collection. The url scheme will work just like it did with CM1.