[CM2] Introduce facets
An interactive rebase of months worth of thinking, doing and not doing.

Faceting is similar to the earlier tag indexing. With facets, posts can be indexed by any fields, not just tags.

With this, tag indexing is no longer a special case. There's no need for modelling or rendering tags.

Facets are created based on 'facetKeys' of a contentType. A simple array of keys works.

Now, each collection has a new route: '/by' which is listing all available facet views in the collection.

And each facet view has a route: /by/facet-name. E.g. /by/author, /by/date, /by/tags. In that page, possible values for the facet can be listed.

And each possible value for the facet has the route: /by/facet-name/value. E.g. /by/author/enes, /by/date/1970-11-23, /by/tags/css. There, posts matching the criteria can be shown.

These new render targets should be implemented inside the theme, and that's what should come soon. So far, working with a local laboratory that has a custom theme.

A notable side-effect of turning a field into a facet is that the field becomes an object in the shape of: { value, facetPermalink }. So, if 'date' field is a facet, templates should do: {{date.value}} to display its value and {{date.facetPermalink}} to render a link to the facet value page in the collection. The name 'facetPermalink' is so ugly tho.

If a facet field links to another entry, then the field's value, which is populated with the linked entry, just gets a 'facetPermalink', to make it possible to render links both to the linked entry and to the facet value page.

Facets work across deep category tree. So, all of the mentioned routes exist at each stop.

E.g.
site.com/photos/by/subject,
site.com/photos/astro/by/subject,
site.com/photos/astro/deep-space/by-subject

An 'afterEffects' mechanism is invented for models. Facets needed to be collected after the tree-building and linking are finished. So, all the way from the root, each model does whatever side-effect jobs inside its afterEffects function, and then calls its child models' afterEffects function. This was needed by the facets to work with linked fields.

Not all is perfectly polished yet. So follow-up commits are expected.

wishlist: /blog/by/date/1993/10/27, /blog/by/date/1993/10, /blog/by/date/1993 each listing corresponding entries. Nested facets? And then nested tags would be possible. But how. After a month or two of morning walks, I figured this is a pretty complex thought that involves not-yet-existing structures to be implemented and working properly. Until it becomes the next easy possibility, nested facets can wait.