[CM2] Delete tag model and parseArray helper
Tags being no longer needed, it was ok to nuke tag model.

The only remaining special care for a tag field was inside the post
model, which was parsing a single line of comma separated strings as an array,
using a parseArray function. All those are deleted.

The only other user of the parseArray helper was the collection model,
using it to parse 'facetKeys' in the same way.

With the removal of that helper, all arrays in the front-matter should
now follow the official syntax.

This means an array field, like tags, should look like either one of these:

tags: [tag1, tag2, tag3]

or

tags:
- tag1
- tag2
- tag3