[CM2] Implement array of links in entry fields
This enables linking to multiple entries for the same key from an entry.

For this to work, linking syntax is updated to:

author: +writers/enes

A category slug can also be used:

author: +writers/turkish/enes

Deep categories are not handled.

So, a multiple linking example:

guests: [+people/enes, +people/daft, +people/punk]

TODO: Currently, collecting facets from these arrays are broken and
should be fixed soon with a follow-up.