[CM2] Link back to entries via named props
Content type of an entry can specify through which fields can serve as connection points to entries that link to them.

e.g.

Article's author links to writers/enes, as usual

AND NOW

Person's articles can be a bucket of those Articles.

SO

If the content type of an entry specifies fields that accepts links, those fields will be used during backlinking, instead of pushing to its relations object.

For this to work, the content type should use this syntax for a 'bucket' field:

(a person content type)
articles: +Article:author
parents: [+Person:sons, +Person:daughters]
photos: +:photographer

in other words, the value of a field should specify the linking entry's field name and optionally its content type. And multiple of these can be provided as an array.

This is a first draft and it needs more thoughts. Currently, the
previous behaviour of pushing into target entry's relations object is
preserved for when the target entry doesn't have a content type or when
it doesn't specify anything related to linking or otherwise no buckets
are found in target's content type.