[CM2] Create theme/common/partials2
It contains new page partials like collection and facet.

So, if compilerVersion is 2:
- theme settings is overridden as zen
- partials2 is copied and registered

Also, introduce a serialize function in facet and collection.

- Previously facet model was using a 'presentationFacets' for making
templates' job easy. Now that logic is moved to a serialize function and
it is exported as a model method.
- Collection's serialize just maps its facets to their serialized
versions
- Although feels wrong, homepage model now requires collection model to
be able to serialize it for the homepage template to access collections'
facets in the same way as collection page does.

All this definitely inspires a serialization step of contentModel before
rendering.