[CM2] Render [collectionName].json
Render it to the directory containing the collection to get an 'add
.json at the end and get data' effect.
i.e. /articles is the html, /articles.json is the json

The JSON includes only the posts of a collection.

Create a safeStringify helper that stringifies a JSON even if it has
circular references in it. It omits the undesired keys (using omit) and
maps duplicate objects to a few select keys of them (using stub)

Also add a renderer.createFile for generic file outputs