Dirty fix subpage permalink and outputPath
At the time of computing permalink and outputPath, which happens inside
super(), the subpage slug does not yet exist, because parseTextEntry is
not called yet.

So... Just call getPermalink and getOutputPath again inside constructor
with the slug and hasIndex retrieved from parseTextEntry. This needs
some design changes, but not right now