Extend models/post from ContentModelNode
When collection wants to create post from data, instead of normal fsNode
parsing, it passes the data in place of fsNode. Then, post checks if
'fsNode' has stats?.birthtime, which it only has if it's actually a file
system node and not data coming from [collectionName].json.

After determining what kind of data node it deals with, the post passes
a flag named isFlatData to its buddy parseTextEntry, which now has a way
of dealing with those kinds of 'flat data'. Basically it just returns
the data itself + parsing the data.content, if data has a content field.

matchers is updated to add post matchers and modernize subpage matchers'
naming. A pattern now starts to emerge. Models have to keep their own
copy of matchers for their child models. A way to re-use matchers from
one place is needed.