Modernize models/category for its post and attachment usages
Also have at least an empty array of facets in the category.

If not, when the collection doesn't have facets, category wouldn't have
facetKeys (inherits from context), and that causes category to not
collectFacets, which was the only place where 'category.facets ='
happens. Then, serialize function tries to map category.facets,
regardless of whether it actually has facets. And boom.