Fix entryAlias not working in uncategorized post
In ContentModelEntryNode, use getIndexFile template method to set
this.indexFile (and not this.subtree.indexFile). Once the indexFile is
set and its content and frontmatter is parsed, continue calling other
template methods, importantly, getSubtreeMatchers.

After switching to use matcha expressions in matchers, they got out of
their function wrappers. So, if they wanted to refer to this properties
coming from front-matter, getSubtreeMatchers must have been called in
super after indexFile is parsed. This change does just that.

Setting this.indexFile instead of this.subtree.indexFile is for the
coherence. If indexFile remained under 'subtree', then it would be
confusing why its matcher is not in the getSubtreeMatchers. It feels
right to have it directly in 'this'.

Also add a missing edge-case to e2e-magazine: Uncategorized foldered
post with an aliased indexFile. Lack of this test so far was what
allowed this bug to hide so far.