Automate childNode matching and parsing in ContentModelEntryNode
Introduce a getSubtreeConfig method, in which all necessities of
matching childNodes and building subtree out of them are defined.

For each key in subtree, a definition like this is added to the config
array:

{
  matcher: matcha;
  key?: string;
  settings?: object;
  model?: ContentModelNode constructor;
  singular?: boolean (false);
  sideEffect?: function;
}

Then parseSubtree template method is actually implemented in the
ContentModelEntryNode. It loops over childNodes and subtree configs.

For everything to look and feel smooth some additional changes are made:

- ContentModel does not pass a contentType property to collection any
more. Instead, collection itself computes it based on the
settings.contentTypes in its getSchema method
- Compiler passes the { children: fileSystemTree } to the contentModel
- static draftCheck is now a class method of ContentModelEntryNode
- A getChildContext template method is added for models to be able to
customize the context they pass to their childNodes. E.g. ContentModel,
AssetsDirectory, PagesDirectory