Introduce exclusive one-to-one linking
Just like how a post can link to an author as a singular link, now the
author can link back to a single post.

Schema's linked-field definitions now have to use array notation to
allow linking back to many entries. If it is in the form of:

soulmate: +Person/foo

then the field will be populated with a single entry.

This is a preparation step for making linking more abstract. Make the
entry schema responsible for the linked-field's type.

Previously all backlinks were collected in 'bucket' fields and schema
had no power to change that. Now, it has.