Link nodes using a flat list of all nodes
Instead of traversing the contentModel's collections and categories,
flatten all nodes (including collections, categories, subpages and
posts) and link them.

Now, the linking logic is not coupled with contentModel structure.

Also linking works in a more flexible way. Linking to a node can be as
simple as:

myNode: +my-node

given my-node is the slug of the node to link to, and there are no other
nodes with the same slug. If there are multiple nodes with the same
slug, then a path to node should be specified, like this:

myNode: +this/way/my-node

and 'this' can be a category of any depth, as well as a collection.
Otherwise link is not added.

New e2e tests needed to verify all this.

Currently, if multiple nodes in a category share the same slug, and if
the path specified is just the category slug, then the first match is
linked. Maybe this is inconsistent with the result of other same-slug
scenario (when no path is specified).