Extract parsing linked fields into parseTextEntry
So the linking logic in the contentModel/index is thinner — more
abstract. An unfortunate result of extraction is that now two places are
digging into entry fields in the same way. Feels like a tooth cavity.
But it is necessary to dig the array field items one by one to be able
to preserve non-linked items in them.